Move the GPU feature detection from gl_state.cpp to gpu_features.cpp
